Added main.rs temp to use it as a cli app
The existing functions have been gathered under the options folder, allowing the requested function to be called directly from the main through selections. In the future, a separate folder will be created for smaller functions that can be shared across the code in the options folder.

main.rs (will be) using /options
options will be using /funcs(?)

For now, no changes have been made except making the functions public; details about the options code will be refactored in the future for cli usage.